# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CampusCoin API
  version: 0.1.0
  description: CampusCoin ("Smart Spending, Student Style") REST API. Money amounts are decimal strings (never floats); errors follow RFC 9457 `application/problem+json`; every route scoped to the caller's own data returns 404 (not 403) for another user's resource.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Short-lived access token (EdDSA JWT) from `POST /auth/login`, sent as `Authorization: Bearer <token>`."
  schemas:
...
    changePasswordSchema:
      type: object
      properties:
        currentPassword:
          type: string
          minLength: 1
          maxLength: 128
        newPassword:
          type: string
          minLength: 10
          maxLength: 128
      required:
        - currentPassword
        - newPassword
      additionalProperties: false
...
    deleteAccountSchema:
      type: object
      properties:
        password:
          type: string
          minLength: 1
          maxLength: 128
        confirm:
          type: string
          enum:
            - DELETE
      required:
        - password
        - confirm
      additionalProperties: false
...
    updateProfileSchema:
      type: object
      properties:
        fullName:
          type: string
        academicYear:
          type:
            - string
            - "null"
          enum:
            - year_1
            - year_2
            - year_3
            - year_4
            - year_5
            - postgraduate
            - other
            - null
        monthlyAllowanceBaseline:
          type:
            - string
            - "null"
          pattern: ^\d{1,12}(\.\d{1,2})?$
        monthlySavingsGoal:
          type:
            - string
            - "null"
          pattern: ^\d{1,12}(\.\d{1,2})?$
        currency:
          type: string
          enum:
            - USD
            - VND
        timezone:
          type: string
        preferences:
          type: object
          properties:
            theme:
              type: string
              enum:
                - light
                - dark
                - system
            fontScale:
              type: number
              minimum: 0.875
              maximum: 1.5
            locale:
              type: string
              enum:
                - en
          additionalProperties: false
        aiOptIn:
          type: boolean
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          description: Machine-readable error slug, e.g. "validation-failed".
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
            required:
              - field
              - message
          description: Present on 422 validation-failed responses, one entry per invalid field.
      required:
        - type
        - title
        - status
      description: RFC 9457 application/problem+json body.
...
paths:
...
  /api/v1/me:
    get:
      tags:
        - Users
      summary: Get the current user's profile.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    patch:
      tags:
        - Users
      summary: Update profile fields (name, preferences).
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/updateProfileSchema"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    delete:
      tags:
        - Users
      summary: Delete the current account (right to erasure) — student-only.
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/deleteAccountSchema"
      responses:
        "202":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/me/password:
    patch:
      tags:
        - Users
      summary: Change password (keeps the current session, revokes the rest).
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/changePasswordSchema"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/me/sessions:
    get:
      tags:
        - Users
      summary: List active sessions (refresh-token families).
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/me/sessions/{id}:
    delete:
      tags:
        - Users
      summary: Revoke one session by id.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/me/export:
    get:
      tags:
        - Users
      summary: Export all personal data (JSON or CSV) — student-only, 3/day.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            enum:
              - json
              - csv
            default: json
          required: false
          name: format
          in: query
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: One or more fields failed validation.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /api/v1/tips:
    get:
      tags:
        - Tips
      summary: List personalised saving tips.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/tips/{id}/pin:
    post:
      tags:
        - Tips
      summary: Pin a tip.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^\d{1,20}$
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/tips/{id}/unpin:
    post:
      tags:
        - Tips
      summary: Unpin a tip.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^\d{1,20}$
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/v1/tips/{id}/dismiss:
    post:
      tags:
        - Tips
      summary: Dismiss a tip.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^\d{1,20}$
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: object
                properties: {}
                additionalProperties: {}
        "401":
          description: Missing/invalid access token.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Resource not found (or owned by another user).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Rate limit exceeded.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...